separate commands and mouse navigation
