+++ src/vs/editor/browser/editor.all.ts
@@ -23,7 +23,8 @@ import 'vs/editor/contrib/dnd/browser/dnd';
 import 'vs/editor/contrib/find/browser/find';
 import 'vs/editor/contrib/folding/browser/folding';
 import 'vs/editor/contrib/format/browser/formatActions';
-import 'vs/editor/contrib/goToDeclaration/browser/goToDeclaration';
+import 'vs/editor/contrib/goToDeclaration/browser/goToDeclarationCommands';
+import 'vs/editor/contrib/goToDeclaration/browser/goToDeclarationMouse';
 import 'vs/editor/contrib/gotoError/browser/gotoError';
 import 'vs/editor/contrib/hover/browser/hover';
 import 'vs/editor/contrib/inPlaceReplace/common/inPlaceReplace';
